# backend/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Japan SSW Platform API",
    "version": "1.0.0",
    "description": "REST API for Japan Specified Skilled Worker (SSW) job platform - connects Filipino jobseekers with Japanese employers",
    "contact": {
      "name": "API Support",
      "email": "support@japanssw.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.japanssw.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token obtained from login"
      }
    },
...
  },
...
  "paths": {
...
    "/api/v1/profile": {
      "get": {
        "summary": "Get user profile",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully"
          }
        }
      },
      "post": {
        "summary": "Create user profile",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "dateOfBirth": {
                    "type": "string",
                    "format": "date"
                  },
                  "nationality": {
                    "type": "string"
                  },
                  "japaneseLevel": {
                    "type": "string",
                    "enum": [
                      "N5",
                      "N4",
                      "N3",
                      "N2",
                      "N1",
                      "Native"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Profile created"
          }
        }
      },
      "put": {
        "summary": "Update user profile",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile updated"
          }
        }
      },
      "delete": {
        "summary": "Delete user profile",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile deleted"
          }
        }
      }
    },
    "/api/v1/profile/education": {
      "post": {
        "summary": "Add education entry",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "school": {
                    "type": "string"
                  },
                  "degree": {
                    "type": "string"
                  },
                  "field": {
                    "type": "string"
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "current": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Education added"
          }
        }
      }
    },
    "/api/v1/profile/education/{edu_id}": {
      "put": {
        "summary": "Update education entry",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "edu_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Education updated"
          }
        }
      },
      "delete": {
        "summary": "Delete education entry",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "edu_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Education deleted"
          }
        }
      }
    },
    "/api/v1/profile/experience": {
      "post": {
        "summary": "Add work experience",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "company": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "current": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Experience added"
          }
        }
      }
    },
    "/api/v1/profile/experience/{exp_id}": {
      "put": {
        "summary": "Update experience entry",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "exp_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Experience updated"
          }
        }
      },
      "delete": {
        "summary": "Delete experience entry",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "exp_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Experience deleted"
          }
        }
      }
    },
    "/api/v1/profile/skills": {
      "put": {
        "summary": "Update skills",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "skills": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "level": {
                          "type": "string",
                          "enum": [
                            "Beginner",
                            "Intermediate",
                            "Advanced",
                            "Expert"
                          ]
                        },
                        "category": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Skills updated"
          }
        }
      }
    },
    "/api/v1/profile/languages": {
      "put": {
        "summary": "Update languages",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Languages updated"
          }
        }
      }
    },
    "/api/v1/profile/certifications": {
      "put": {
        "summary": "Update certifications",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Certifications updated"
          }
        }
      }
    },
    "/api/v1/profile/availability": {
      "put": {
        "summary": "Update availability preferences",
        "tags": [
          "Profile"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Availability updated"
          }
        }
      }
    },
...
  }
}